requirements/dev.txt
-r test.txt

ipython

django-debug-toolbar
django-extensions

pylama==7.0.7
Sphinx==1.3.6
sphinx-intl==0.9.9  # translation tool
